# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Akamai CCM API
  version: 1.0.0
  description: >
    Cloud Compliance Manager API for managing Linode accounts, resources,
    compliance rules, reports, and users.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://<your-host>
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        quarter:
          type: string
          nullable: true
        status:
          type: string
          enum: [generating, ready]
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        snapshot:
          type: object
          nullable: true
...
    RefreshResult:
      type: object
      properties:
        success:
          type: boolean
        accounts_processed:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              account_id:
                type: string
                format: uuid
              sync:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
              eval:
                type: object
                properties:
                  success:
                    type: boolean
              error:
                type: string
                nullable: true
        log:
          type: array
          items:
            type: string
        completed_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        detail:
          type: string
security:
  - bearerAuth: []
paths:
...
  /api/reports:
    get:
      tags: [Reports]
      summary: List reports
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
    post:
      tags: [Reports]
      summary: Create a compliance report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, title, period_start, period_end]
              properties:
                account_id:
                  type: string
                  format: uuid
                title:
                  type: string
                description:
                  type: string
                  default: ''
                period_start:
                  type: string
                  format: date-time
                period_end:
                  type: string
                  format: date-time
                quarter:
                  type: string
      responses:
        '200':
          description: Created report with snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
  /api/reports/{report_id}:
    parameters:
      - name: report_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Reports]
      summary: Get report details
      responses:
        '200':
          description: Full report with snapshot data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '404':
          description: Report not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Reports]
      summary: Delete a report
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  # ─── Events ───────────────────────────────────────────────────────────────
...
  /api/refresh:
    post:
      tags: [Refresh]
      summary: Trigger sync and compliance evaluation
      description: Requires `power_user` or `admin` role. Omit `account_id` to process all accounts.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                account_id:
                  type: string
                  format: uuid
                  nullable: true
                skip_sync:
                  type: boolean
                  default: false
                skip_eval:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Refresh result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResult'
    get:
      tags: [Refresh]
      summary: Trigger refresh via GET (webhook / cron)
      description: >
        Auth is optional — only required if `REFRESH_API_SECRET` is configured on the server.
        Pass the token via `?token=` query parameter or `Authorization: Bearer` header.
      security: []
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          schema:
            type: string
          description: API secret token (alternative to Authorization header)
        - name: skip_sync
          in: query
          schema:
            type: boolean
        - name: skip_eval
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Refresh result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResult'
  /api/refresh/stream:
    get:
      tags: [Refresh]
      summary: Stream refresh progress (SSE)
      description: >
        Returns a Server-Sent Events stream. Requires `power_user` or `admin` role.

        **Event types:**
        - `phase` — Phase started (`sync` or `evaluate`)
        - `sync_done` — Sync completed with resource count
        - `eval_done` — Evaluation phase completed
        - `log` — Log message string
        - `done` — Final summary object
        - `error` — Error message string
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: skip_sync
          in: query
          schema:
            type: boolean
        - name: skip_eval
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string

  # ─── Admin ────────────────────────────────────────────────────────────────
...